update chainlink test
