+++ test/__init__.py
@@ -1,3 +1,4 @@
 from .SeedEmuTestCase import SeedEmuTestCase
 from .internet import *
 from .ethereum import *
+from .chainlink import *
