+++ test/chainlink/POA/emulator-code/test-emulator.py
@@ -7,13 +7,167 @@ import platform
 from typing import List
 
 ###############################################################################
-emu = Emulator()
-base = Base()
+#!/usr/bin/env python3
+# encoding: utf-8
 
-# Load the pre-built components and merge them
-emu.load('./hybrid-internet.bin')
-base = emu.getLayer('Base')
+from seedemu import *
 
+def makeStubAs(emu: Emulator, base: Base, asn: int, exchange: int, hosts_total: int):
+    # Create AS and internal network
+    network = "net0"
+    stub_as = base.createAutonomousSystem(asn)
+    stub_as.createNetwork(network)
+
+    # Create a BGP router
+    # Attach the router to both the internal and external networks
+    router = stub_as.createRouter('router0')
+    router.joinNetwork(network)
+    router.joinNetwork('ix{}'.format(exchange))
+
+    for counter in range(hosts_total):
+       name = 'host_{}'.format(counter)
+       host = stub_as.createHost(name)
+       host.joinNetwork(network)
+
+
+
+hosts_total = 3
+
+
+###############################################################################
+emu     = Emulator()
+base    = Base()
+routing = Routing()
+ebgp    = Ebgp()
+ibgp    = Ibgp()
+ospf    = Ospf()
+web     = WebService()
+ovpn    = OpenVpnRemoteAccessProvider()
+
+
+###############################################################################
+
+ix100 = base.createInternetExchange(100)
+ix101 = base.createInternetExchange(101)
+ix102 = base.createInternetExchange(102)
+ix103 = base.createInternetExchange(103)
+ix104 = base.createInternetExchange(104)
+ix105 = base.createInternetExchange(105)
+
+# Customize names (for visualization purpose)
+ix100.getPeeringLan().setDisplayName('NYC-100')
+ix101.getPeeringLan().setDisplayName('San Jose-101')
+ix102.getPeeringLan().setDisplayName('Chicago-102')
+ix103.getPeeringLan().setDisplayName('Miami-103')
+ix104.getPeeringLan().setDisplayName('Toronto-104')
+ix105.getPeeringLan().setDisplayName('Huston-105')
+
+
+###############################################################################
+# Create Transit Autonomous Systems 
+
+## Tier 1 ASes
+Makers.makeTransitAs(base, 2, [100, 101, 102, 105], 
+       [(100, 101), (101, 102), (100, 105)] 
+)
+
+Makers.makeTransitAs(base, 3, [100, 103, 104, 105], 
+       [(100, 103), (100, 105), (103, 105), (103, 104)]
+)
+
+Makers.makeTransitAs(base, 4, [100, 102, 104], 
+       [(100, 104), (102, 104)]
+)
+# This transit is for lab exercise. Its setup is incomplete
+Makers.makeTransitAs(base, 5, [101, 103, 105], 
+       [(101, 103), (103, 105)]
+)
+
+## Tier 2 ASes
+Makers.makeTransitAs(base, 11, [102, 105], [(102, 105)])
+Makers.makeTransitAs(base, 12, [101, 104], [(101, 104)])
+
+
+###############################################################################
+# Create single-homed stub ASes. "None" means create a host only 
+
+makeStubAs(emu, base, 150, 100, hosts_total)
+makeStubAs(emu, base, 151, 100, hosts_total)
+
+makeStubAs(emu, base, 152, 101, hosts_total)
+makeStubAs(emu, base, 153, 101, hosts_total)
+
+makeStubAs(emu, base, 154, 102, hosts_total)
+makeStubAs(emu, base, 155, 102, hosts_total)
+makeStubAs(emu, base, 156, 102, hosts_total)
+
+makeStubAs(emu, base, 160, 103, hosts_total)
+makeStubAs(emu, base, 161, 103, hosts_total)
+makeStubAs(emu, base, 162, 103, hosts_total)
+
+makeStubAs(emu, base, 163, 104, hosts_total)
+makeStubAs(emu, base, 164, 104, hosts_total)
+
+makeStubAs(emu, base, 170, 105, hosts_total)
+makeStubAs(emu, base, 171, 105, hosts_total)
+
+# Create real-world AS.
+# AS11872 is the Syracuse University's autonomous system
+
+as11872 = base.createAutonomousSystem(11872)
+as11872.createRealWorldRouter('rw-11872-syr').joinNetwork('ix102', '10.102.0.118')
+
+
+###############################################################################
+# Create hybrid AS.
+# AS99999 is the emulator's autonomous system that routes the traffics to the real-world internet
+as99999 = base.createAutonomousSystem(99999)
+as99999.createRealWorldRouter('rw-real-world', prefixes=['0.0.0.0/1', '128.0.0.0/1']).joinNetwork('ix100', '10.100.0.99')
+###############################################################################
+
+
+###############################################################################
+# Peering via RS (route server). The default peering mode for RS is PeerRelationship.Peer, 
+# which means each AS will only export its customers and their own prefixes. 
+# We will use this peering relationship to peer all the ASes in an IX.
+# None of them will provide transit service for others. 
+
+ebgp.addRsPeers(100, [2, 3, 4])
+ebgp.addRsPeers(102, [2, 4])
+ebgp.addRsPeers(104, [3, 4])
+ebgp.addRsPeers(105, [2, 3])
+
+# To buy transit services from another autonomous system, 
+# we will use private peering  
+
+ebgp.addPrivatePeerings(100, [2],  [150, 151], PeerRelationship.Provider)
+ebgp.addPrivatePeerings(100, [3],  [150, 99999], PeerRelationship.Provider)
+
+ebgp.addPrivatePeerings(101, [2],  [12], PeerRelationship.Provider)
+ebgp.addPrivatePeerings(101, [12], [152, 153], PeerRelationship.Provider)
+
+ebgp.addPrivatePeerings(102, [2, 4],  [11, 154], PeerRelationship.Provider)
+ebgp.addPrivatePeerings(102, [11], [154, 11872], PeerRelationship.Provider)
+
+ebgp.addPrivatePeerings(103, [3],  [160, 161, 162 ], PeerRelationship.Provider)
+
+ebgp.addPrivatePeerings(104, [3, 4], [12], PeerRelationship.Provider)
+ebgp.addPrivatePeerings(104, [4],  [163], PeerRelationship.Provider)
+ebgp.addPrivatePeerings(104, [12], [164], PeerRelationship.Provider)
+
+ebgp.addPrivatePeerings(105, [3],  [11, 170], PeerRelationship.Provider)
+ebgp.addPrivatePeerings(105, [11], [171], PeerRelationship.Provider)
+
+
+###############################################################################
+
+# Add layers to the emulator
+emu.addLayer(base)
+emu.addLayer(routing)
+emu.addLayer(ebgp)
+emu.addLayer(ibgp)
+emu.addLayer(ospf)
+emu.addLayer(web)
 
 ###############################################################################
 # Create the Ethereum layer
@@ -73,6 +227,7 @@ faucet:FaucetServer = blockchain.createFaucetServer(vnode='faucet',
                                                      max_fund_amount=20)
 
 emu.addBinding(Binding('faucet', filter=Filter(asn=150, nodeName='host_2')))
+emu.getVirtualNode('faucet').setDisplayName('FaucetServer')
 
 # Create the Chainlink Init server
 chainlink = ChainlinkService()
